refactor(gateway): split runtime responsibilities
