+++ crates/omniinfer-cli/src/gateway/mod.rs
@@ -176,600 +176,9 @@ async fn proxy_request_inner(
     ))
 }
 
-async fn should_handle_rust_endpoint(state: &GatewayState, method: &Method, path: &str) -> bool {
-    match (method, path) {
-        (
-            &Method::GET,
-            "/health"
-            | "/omni/state"
-            | "/omni/backends"
-            | "/omni/thinking"
-            | "/omni/models"
-            | "/omni/gpus"
-            | "/omni/loaded-models",
-        ) => true,
-        (&Method::GET, "/omni/backend/props") => true,
-        (&Method::GET, "/omni/public-models") => true,
-        (&Method::GET, path) if request_history_path(path) => true,
-        (&Method::GET, "/omni/supported-models" | "/omni/supported-models/best" | "/v1/models") => {
-            true
-        }
-        (&Method::POST, "/omni/shutdown") => true,
-        (
-            &Method::POST,
-            "/omni/backend/select"
-            | "/omni/backend/stop"
-            | "/omni/model/clear-selection"
-            | "/omni/model/select"
-            | "/omni/model/load"
-            | "/omni/model/unload",
-        ) => true,
-        (&Method::POST, "/omni/thinking/select") => true,
-        (&Method::POST, "/v1/chat/completions" | "/v1/messages") => true,
-        (
-            &Method::POST,
-            "/tokenize" | "/detokenize" | "/omni/tokenize" | "/omni/detokenize"
-            | "/omni/cache/clear",
-        ) => state.runtime.lock().await.has_loaded_runtime(),
-        _ => false,
-    }
-}
+mod management;
 
-async fn try_handle_rust_endpoint(
-    state: &GatewayState,
-    path: &str,
-    auth: GatewayAuthDecision,
-    request: Request<Body>,
-) -> Result<Option<Response<Body>>> {
-    match (request.method(), path) {
-        (&Method::GET, "/health") => {
-            let deep = request
-                .uri()
-                .query()
-                .map(|query| query.contains("deep=true") || query.contains("deep=1"))
-                .unwrap_or(false);
-            let snapshot = state.runtime.lock().await.snapshot();
-            let mut payload = json!({
-                "status": "ok",
-                "omni": snapshot,
-            });
-            if deep {
-                payload["backend_health"] = backend_health(&snapshot);
-            }
-            Ok(Some(json_response(StatusCode::OK, payload)))
-        }
-        (&Method::GET, "/omni/state") => {
-            let mut payload = state.runtime.lock().await.snapshot();
-            payload["available_backends"] =
-                BackendRegistry::load_current().api_payload(BackendScope::All)["data"].clone();
-            Ok(Some(json_response(StatusCode::OK, payload)))
-        }
-        (&Method::GET, "/omni/backends") => {
-            let scope = request
-                .uri()
-                .query()
-                .and_then(|query| {
-                    query.split('&').find_map(|part| {
-                        let (key, value) = part.split_once('=')?;
-                        (key == "scope").then_some(value)
-                    })
-                })
-                .unwrap_or("installed");
-            let scope = match scope {
-                "installed" => BackendScope::Installed,
-                "compatible" => BackendScope::Compatible,
-                "all" => BackendScope::All,
-                other => {
-                    return Ok(Some(json_response(
-                        StatusCode::BAD_REQUEST,
-                        json!({"error": {"message": format!("invalid scope: {other}. Must be one of: installed, compatible, all")}}),
-                    )));
-                }
-            };
-            Ok(Some(json_response(
-                StatusCode::OK,
-                BackendRegistry::load_current().api_payload(scope),
-            )))
-        }
-        (&Method::GET, "/omni/thinking") => Ok(Some(json_response(
-            StatusCode::OK,
-            json!({"default_enabled": default_thinking_enabled()}),
-        ))),
-        (&Method::GET, "/omni/backend/props") => {
-            let target = state.runtime.lock().await.proxy_base_for_model(None);
-            let Some(target) = target else {
-                return Ok(Some(json_response(StatusCode::OK, json!({}))));
-            };
-            let response = proxy_get_to_runtime(&state.client, &format!("{target}/props")).await?;
-            Ok(Some(response))
-        }
-        (&Method::GET, "/omni/loaded-models") => {
-            let payload = state.runtime.lock().await.loaded_models_payload();
-            Ok(Some(json_response(StatusCode::OK, payload)))
-        }
-        (&Method::GET, "/omni/gpus") => {
-            let loaded = state.runtime.lock().await.loaded_runtime_summaries();
-            match query_nvidia_smi_gpu_status(&loaded) {
-                Ok(devices) => Ok(Some(json_response(
-                    StatusCode::OK,
-                    gpu_status_payload(&devices),
-                ))),
-                Err(error) => Ok(Some(json_response(
-                    StatusCode::SERVICE_UNAVAILABLE,
-                    json!({"error": {"message": error.to_string()}}),
-                ))),
-            }
-        }
-        (&Method::GET, "/omni/models") => Ok(Some(json_response(
-            StatusCode::GONE,
-            json!({"error": {"message": "GET /omni/models has been deprecated and is no longer maintained"}}),
-        ))),
-        (&Method::GET, "/omni/public-models") => {
-            match public_models::list_public_models(state.public_model_root.as_deref()) {
-                Ok(entries) => Ok(Some(json_response(
-                    StatusCode::OK,
-                    public_models::public_models_payload(&entries),
-                ))),
-                Err(error) => Ok(Some(json_response(
-                    public_model_error_status(&error),
-                    json!({"error": {"message": error.to_string()}}),
-                ))),
-            }
-        }
-        (&Method::GET, path) if request_history_path(path) => {
-            if auth.admin_id.is_none() {
-                return Ok(Some(json_response(
-                    StatusCode::FORBIDDEN,
-                    json!({"error": {"message": "request history requires an admin key"}}),
-                )));
-            }
-            if let Some(id) = path.strip_prefix("/omni/request-history/") {
-                let history_dir = state.request_history_dir.clone();
-                let id = id.to_string();
-                let lookup_id = id.clone();
-                let result = tokio::task::spawn_blocking(move || {
-                    request_history::get_record(&history_dir, &lookup_id)
-                })
-                .await??;
-                return Ok(Some(match result {
-                    Some(entry) => json_response(StatusCode::OK, entry),
-                    None => json_response(
-                        StatusCode::NOT_FOUND,
-                        json!({"error": {"message": format!("request history entry not found: {id}")}}),
-                    ),
-                }));
-            }
-            let query = query_from_pairs(query_pairs(request.uri()));
-            let history_dir = state.request_history_dir.clone();
-            let payload = tokio::task::spawn_blocking(move || {
-                request_history::query_records(&history_dir, query)
-            })
-            .await??;
-            Ok(Some(json_response(StatusCode::OK, payload)))
-        }
-        (&Method::GET, "/omni/supported-models") => {
-            let system = query_value(request.uri(), "system").unwrap_or_else(current_system_name);
-            match model_catalog::list_supported_models(&system) {
-                Ok(payload) => Ok(Some(json_response(StatusCode::OK, payload))),
-                Err(error) => Ok(Some(json_response(
-                    StatusCode::BAD_REQUEST,
-                    json!({"error": {"message": error.to_string()}}),
-                ))),
-            }
-        }
-        (&Method::GET, "/omni/supported-models/best") => {
-            let system = query_value(request.uri(), "system").unwrap_or_else(current_system_name);
-            match model_catalog::list_supported_models_best(&system) {
-                Ok(payload) => Ok(Some(json_response(StatusCode::OK, payload))),
-                Err(error) => Ok(Some(json_response(
-                    StatusCode::BAD_REQUEST,
-                    json!({"error": {"message": error.to_string()}}),
-                ))),
-            }
-        }
-        (&Method::GET, "/v1/models") => {
-            let loaded = state.runtime.lock().await.loaded_models_payload();
-            let data = loaded
-                .get("data")
-                .and_then(Value::as_array)
-                .cloned()
-                .unwrap_or_default()
-                .into_iter()
-                .map(|item| {
-                    let id = item
-                        .get("id")
-                        .and_then(Value::as_str)
-                        .unwrap_or("omniinfer")
-                        .to_string();
-                    json!({
-                        "id": id,
-                        "object": "model",
-                        "created": 0,
-                        "owned_by": "omniinfer",
-                        "permission": [],
-                        "root": id,
-                        "parent": null,
-                    })
-                })
-                .collect::<Vec<_>>();
-            Ok(Some(json_response(
-                StatusCode::OK,
-                json!({"object": "list", "data": data}),
-            )))
-        }
-        (&Method::POST, "/omni/backend/select") => {
-            let body = request.into_body().collect().await?.to_bytes();
-            let payload: Value = serde_json::from_slice(&body)?;
-            let Some(backend_id) = payload
-                .get("backend")
-                .and_then(Value::as_str)
-                .filter(|value| !value.trim().is_empty())
-            else {
-                return Ok(Some(json_response(
-                    StatusCode::BAD_REQUEST,
-                    json!({"error": {"message": "field 'backend' is required"}}),
-                )));
-            };
-            let result = state.runtime.lock().await.select_backend(backend_id);
-            Ok(Some(json_response(StatusCode::OK, result?)))
-        }
-        (&Method::POST, "/omni/backend/stop") => {
-            let result = tokio::task::spawn_blocking({
-                let runtime = Arc::clone(&state.runtime);
-                move || {
-                    let handle = tokio::runtime::Handle::current();
-                    handle.block_on(async move { runtime.lock().await.stop_runtime() })
-                }
-            })
-            .await??;
-            Ok(Some(json_response(StatusCode::OK, result)))
-        }
-        (&Method::POST, "/omni/model/clear-selection") => {
-            let mut runtime = state.runtime.lock().await;
-            let selection_cleared = local_state::clear_selected_model()?;
-            let snapshot = runtime.snapshot();
-            Ok(Some(json_response(
-                StatusCode::OK,
-                json!({
-                    "ok": true,
-                    "selection_cleared": selection_cleared,
-                    "backend_ready": snapshot["backend_ready"],
-                    "current_model": snapshot["model"],
-                    "restore_selection": snapshot["restore_selection"],
-                    "restore_status": snapshot["restore_status"],
-                    "restore_completed": snapshot["restore_completed"],
-                }),
-            )))
-        }
-        (&Method::POST, "/omni/shutdown") => {
-            let result = tokio::task::spawn_blocking({
-                let runtime = Arc::clone(&state.runtime);
-                move || {
-                    let handle = tokio::runtime::Handle::current();
-                    handle.block_on(async move { runtime.lock().await.stop_runtime() })
-                }
-            })
-            .await??;
-            if let Some(sender) = state.shutdown.lock().await.take() {
-                let _ = sender.send(());
-            }
-            Ok(Some(json_response(
-                StatusCode::OK,
-                json!({"ok": true, "runtime": result}),
-            )))
-        }
-        (&Method::POST, "/omni/thinking/select") => {
-            let body = request.into_body().collect().await?.to_bytes();
-            let payload: Value = serde_json::from_slice(&body)?;
-            let raw_enabled = payload.get("enabled").or_else(|| payload.get("think"));
-            let Some(raw_enabled) = raw_enabled else {
-                return Ok(Some(json_response(
-                    StatusCode::BAD_REQUEST,
-                    json!({"error": {"message": "field 'enabled' is required"}}),
-                )));
-            };
-            let enabled = match omniinfer_core::request_normalization::parse_boolish(raw_enabled) {
-                Ok(enabled) => enabled,
-                Err(error) => {
-                    return Ok(Some(json_response(
-                        StatusCode::BAD_REQUEST,
-                        json!({"error": {"message": error.to_string()}}),
-                    )));
-                }
-            };
-            local_state::save_default_thinking(enabled)?;
-            Ok(Some(json_response(
-                StatusCode::OK,
-                json!({"ok": true, "default_enabled": enabled}),
-            )))
-        }
-        (&Method::POST, "/omni/model/select" | "/omni/model/load") => {
-            let body = request.into_body().collect().await?.to_bytes();
-            let mut payload: Value = serde_json::from_slice(&body)?;
-            if payload
-                .get("request_defaults")
-                .is_some_and(|defaults| !defaults.is_object())
-            {
-                return Ok(Some(json_response(
-                    StatusCode::BAD_REQUEST,
-                    json!({"error": {"message": "request_defaults must be an object"}}),
-                )));
-            }
-            if let Err(error) = normalize_public_model_select(&mut payload, state, auth.remote) {
-                return Ok(Some(json_response(
-                    public_model_error_status(&error),
-                    json!({"error": {"message": error.to_string()}}),
-                )));
-            }
-            {
-                let runtime = state.runtime.lock().await;
-                let requested_backend = runtime.resolve_requested_backend(&payload)?;
-                let registry = BackendRegistry::load_current();
-                let backend = registry
-                    .get(&requested_backend)
-                    .ok_or_else(|| anyhow::anyhow!("unsupported backend: {requested_backend}"))?;
-                if backend.runtime_mode == "embedded" {
-                    return Ok(Some(json_response(
-                        StatusCode::BAD_REQUEST,
-                        json!({"error": {"message": format!("{} is an embedded backend. Python control-plane fallback has been removed; use an external-server backend or a backend adapter service.", backend.id)}}),
-                    )));
-                }
-            };
-            let backend_host = state.backend_host.clone();
-            let runtime_startup_timeout = state.runtime_startup_timeout;
-            let runtime = Arc::clone(&state.runtime);
-            let outcome = tokio::task::spawn_blocking(move || {
-                let handle = tokio::runtime::Handle::current();
-                handle.block_on(async move {
-                    runtime.lock().await.load_model(
-                        payload,
-                        backend_host,
-                        runtime_startup_timeout,
-                        auth.admin_id.clone(),
-                    )
-                })
-            })
-            .await??;
-            let (status, result) = match outcome {
-                LoadModelOutcome::Success(result) => (StatusCode::OK, result),
-                LoadModelOutcome::ReloadRequired(result) => (StatusCode::CONFLICT, result),
-            };
-            Ok(Some(json_response(status, result)))
-        }
-        (&Method::POST, "/omni/model/unload") => {
-            let body = request.into_body().collect().await?.to_bytes();
-            let payload: Value = serde_json::from_slice(&body)?;
-            let Some(model) = payload
-                .get("model")
-                .and_then(Value::as_str)
-                .filter(|value| !value.trim().is_empty())
-            else {
-                return Ok(Some(json_response(
-                    StatusCode::BAD_REQUEST,
-                    json!({"error": {"message": "field 'model' is required"}}),
-                )));
-            };
-            match state
-                .runtime
-                .lock()
-                .await
-                .unload_model(model, auth.admin_id.as_deref())
-            {
-                Ok(result) => Ok(Some(json_response(StatusCode::OK, result))),
-                Err(error) => Ok(Some(json_response(
-                    StatusCode::FORBIDDEN,
-                    json!({"error": {"message": error.to_string()}}),
-                ))),
-            }
-        }
-        (&Method::POST, "/v1/chat/completions") => {
-            let body = request.into_body().collect().await?.to_bytes();
-            let raw_payload: Value = serde_json::from_slice(&body)?;
-            let requested_model = raw_payload
-                .get("model")
-                .and_then(Value::as_str)
-                .map(str::to_string);
-            let target = {
-                let mut runtime = state.runtime.lock().await;
-                runtime.proxy_target_for_model(requested_model.as_deref())
-            };
-            let Some(target) = target else {
-                let message = requested_model
-                    .as_deref()
-                    .map(|model| format!("model is not loaded: {model}"))
-                    .unwrap_or_else(|| "no model is loaded".to_string());
-                return Ok(Some(json_response(
-                    if requested_model.is_some() {
-                        StatusCode::NOT_FOUND
-                    } else {
-                        StatusCode::SERVICE_UNAVAILABLE
-                    },
-                    json!({"error": {"message": message}}),
-                )));
-            };
-            let mut normalized_payload = match normalize_chat_request_with_defaults(
-                raw_payload.clone(),
-                &target.request_defaults,
-                false,
-            ) {
-                Ok(payload) => payload,
-                Err(error) => {
-                    return Ok(Some(json_response(
-                        StatusCode::BAD_REQUEST,
-                        json!({"error": {"message": error.to_string()}}),
-                    )));
-                }
-            };
-            let Some(base_url) = target.base_url.as_deref() else {
-                return Ok(Some(backend_protocol_not_supported(
-                    &target,
-                    "/v1/chat/completions",
-                )));
-            };
-            let response_model = requested_model
-                .clone()
-                .unwrap_or_else(|| "omniinfer".to_string());
-            let stream_requested = normalized_payload
-                .payload
-                .get("stream")
-                .and_then(Value::as_bool)
-                .unwrap_or(false);
-            apply_proxy_model(&mut normalized_payload.payload, target.model.as_deref());
-            let started_at = Instant::now();
-            let history_context = StreamHistoryContext {
-                state: state.clone(),
-                admin_id: auth.admin_id.clone(),
-                auth_kind: auth_kind(&auth),
-                method: "POST".to_string(),
-                path: "/v1/chat/completions".to_string(),
-                model: requested_model.clone(),
-                backend: Some(target.backend_id.clone()),
-                request: raw_payload.clone(),
-                response_model: response_model.clone(),
-                started_at,
-            };
-            let (response, captured_response, status, history_deferred) =
-                if should_proxy_vllm_nonstream_via_stream(&target.backend_id, stream_requested) {
-                    let (payload, status) = proxy_openai_nonstream_via_stream(
-                        &state.client,
-                        &format!("{base_url}/v1/chat/completions"),
-                        normalized_payload.payload,
-                        &response_model,
-                    )
-                    .await?;
-                    (
-                        json_response(status, payload.clone()),
-                        Some(payload),
-                        status,
-                        false,
-                    )
-                } else {
-                    proxy_openai_chat_to_runtime(
-                        &state.client,
-                        &format!("{base_url}/v1/chat/completions"),
-                        HyperBytes::from(serde_json::to_vec(&normalized_payload.payload)?),
-                        Some(history_context.clone()),
-                    )
-                    .await?
-                };
-            if !history_deferred {
-                record_request_history(
-                    &state,
-                    RequestHistoryRecord {
-                        admin_id: history_context.admin_id,
-                        auth_kind: history_context.auth_kind,
-                        method: history_context.method,
-                        path: history_context.path,
-                        model: history_context.model,
-                        backend: history_context.backend,
-                        status: status.as_u16(),
-                        latency_ms: duration_ms(started_at.elapsed()),
-                        usage: captured_response
-                            .as_ref()
-                            .and_then(|payload| payload.get("usage").cloned()),
-                        metrics: captured_response
-                            .as_ref()
-                            .and_then(|payload| payload.get("omniinfer_metrics").cloned()),
-                        request: raw_payload,
-                        response: captured_response,
-                        error: (status.as_u16() >= 400)
-                            .then(|| format!("HTTP {}", status.as_u16())),
-                    },
-                );
-            }
-            Ok(Some(response))
-        }
-        (&Method::POST, "/tokenize" | "/detokenize" | "/omni/tokenize" | "/omni/detokenize") => {
-            let body = request.into_body().collect().await?.to_bytes();
-            let operation = if path.ends_with("detokenize") {
-                "detokenize"
-            } else {
-                "tokenize"
-            };
-            let target = state.runtime.lock().await.proxy_base_for_model(None);
-            let Some(target) = target else {
-                return Ok(None);
-            };
-            let response =
-                proxy_body_to_runtime(&state.client, &format!("{target}/{operation}"), body)
-                    .await?;
-            Ok(Some(response))
-        }
-        (&Method::POST, "/omni/cache/clear") => {
-            let target = state.runtime.lock().await.proxy_base_for_model(None);
-            let Some(target) = target else {
-                return Ok(None);
-            };
-            let response = clear_runtime_cache(&state.client, &target).await?;
-            Ok(Some(response))
-        }
-        (&Method::POST, "/v1/messages") => {
-            let body = request.into_body().collect().await?.to_bytes();
-            let payload: Value = serde_json::from_slice(&body)?;
-            let messages = payload.get("messages").and_then(Value::as_array);
-            if !messages.is_some_and(|messages| !messages.is_empty()) {
-                return Ok(Some(json_response(
-                    StatusCode::BAD_REQUEST,
-                    json!({"error": {"type": "invalid_request_error", "message": "messages is required"}}),
-                )));
-            }
-            let response_model = payload
-                .get("model")
-                .and_then(Value::as_str)
-                .map(str::to_string);
-            let openai_payload = anthropic_request_to_openai(&payload);
-            let mut target = {
-                let mut runtime = state.runtime.lock().await;
-                runtime.proxy_target_for_model(response_model.as_deref())
-            };
-            if target.is_none() && response_model.is_some() {
-                target = state.runtime.lock().await.proxy_target_for_model(None);
-            }
-            let Some(target) = target else {
-                return Ok(Some(json_response(
-                    StatusCode::SERVICE_UNAVAILABLE,
-                    json!({"error": {"message": "no model is loaded"}}),
-                )));
-            };
-            let mut normalized = match normalize_chat_request_with_defaults(
-                openai_payload,
-                &target.request_defaults,
-                false,
-            ) {
-                Ok(payload) => payload,
-                Err(error) => {
-                    return Ok(Some(json_response(
-                        StatusCode::BAD_REQUEST,
-                        json!({"error": {"message": error.to_string()}}),
-                    )));
-                }
-            };
-            let Some(base_url) = target.base_url.as_deref() else {
-                return Ok(Some(backend_protocol_not_supported(
-                    &target,
-                    "/v1/messages",
-                )));
-            };
-            let response_model = response_model.unwrap_or_else(|| "omniinfer".to_string());
-            apply_proxy_model(&mut normalized.payload, target.model.as_deref());
-            let response = proxy_anthropic_to_runtime(
-                &state.client,
-                &format!("{base_url}/v1/chat/completions"),
-                HyperBytes::from(serde_json::to_vec(&normalized.payload)?),
-                &response_model,
-                normalized
-                    .payload
-                    .get("stream")
-                    .and_then(Value::as_bool)
-                    .unwrap_or(false),
-            )
-            .await?;
-            Ok(Some(response))
-        }
-        _ => Ok(None),
-    }
-}
-
+use management::{should_handle_rust_endpoint, try_handle_rust_endpoint};
 async fn proxy_get_to_runtime(
     client: &Client<HttpConnector, Full<HyperBytes>>,
     uri: &str,
@@ -844,480 +253,12 @@ async fn proxy_openai_chat_to_runtime(
     Ok((response, captured, status, false))
 }
 
-#[derive(Clone)]
-struct StreamHistoryContext {
-    state: GatewayState,
-    admin_id: Option<String>,
-    auth_kind: String,
-    method: String,
-    path: String,
-    model: Option<String>,
-    backend: Option<String>,
-    request: Value,
-    response_model: String,
-    started_at: Instant,
-}
+mod openai_stream;
 
-fn stream_openai_chat_with_history(
-    upstream_body: hyper::body::Incoming,
-    builder: axum::http::response::Builder,
-    context: StreamHistoryContext,
-    status: StatusCode,
-) -> Result<Response<Body>> {
-    let (tx, rx) = mpsc::channel::<Result<HyperBytes, std::io::Error>>(16);
-    tokio::spawn(async move {
-        let mut body = Body::new(upstream_body);
-        let mut buffered = Vec::<u8>::new();
-        let mut aggregate = OpenAiStreamAggregate::new(
-            &context.response_model,
-            context.started_at,
-            "stream_passthrough",
-        );
-        let mut error = None::<String>;
-        while let Some(frame) = body.frame().await {
-            let frame = match frame {
-                Ok(frame) => frame,
-                Err(frame_error) => {
-                    let message = frame_error.to_string();
-                    let _ = tx.send(Err(std::io::Error::other(message.clone()))).await;
-                    error = Some(format!("upstream stream error: {message}"));
-                    break;
-                }
-            };
-            let Some(data) = frame.data_ref() else {
-                continue;
-            };
-            let chunk = HyperBytes::copy_from_slice(data);
-            if tx.send(Ok(chunk.clone())).await.is_err() {
-                error = Some("client disconnected while streaming response".to_string());
-                break;
-            }
-            buffered.extend_from_slice(&chunk);
-            while let Some(index) = buffered.windows(2).position(|window| window == b"\n\n") {
-                let event = buffered.drain(..index + 2).collect::<Vec<_>>();
-                aggregate.process_sse_bytes(&event);
-            }
-        }
-        if error.is_none() && !buffered.is_empty() {
-            aggregate.process_sse_bytes(&buffered);
-        }
-        let payload = aggregate.finish();
-        let record_error =
-            error.or_else(|| (status.as_u16() >= 400).then(|| format!("HTTP {}", status.as_u16())));
-        record_request_history(
-            &context.state,
-            RequestHistoryRecord {
-                admin_id: context.admin_id,
-                auth_kind: context.auth_kind,
-                method: context.method,
-                path: context.path,
-                model: context.model,
-                backend: context.backend,
-                status: status.as_u16(),
-                latency_ms: duration_ms(context.started_at.elapsed()),
-                usage: payload.get("usage").cloned(),
-                metrics: payload.get("omniinfer_metrics").cloned(),
-                request: context.request,
-                response: Some(payload),
-                error: record_error,
-            },
-        );
-    });
-    let mut response = builder.body(Body::from_stream(ReceiverStream::new(rx)))?;
-    add_cors_headers(response.headers_mut());
-    Ok(response)
-}
-
-fn should_proxy_vllm_nonstream_via_stream(backend_id: &str, stream_requested: bool) -> bool {
-    !stream_requested
-        && backend_id.starts_with("vllm")
-        && env_flag_enabled("OMNIINFER_VLLM_NONSTREAM_VIA_STREAM", true)
-}
-
-fn env_flag_enabled(name: &str, default: bool) -> bool {
-    std::env::var(name)
-        .map(|value| {
-            matches!(
-                value.trim().to_ascii_lowercase().as_str(),
-                "1" | "true" | "yes" | "on"
-            )
-        })
-        .unwrap_or(default)
-}
-
-async fn proxy_openai_nonstream_via_stream(
-    client: &Client<HttpConnector, Full<HyperBytes>>,
-    uri: &str,
-    mut payload: Value,
-    response_model: &str,
-) -> Result<(Value, StatusCode)> {
-    payload["stream"] = json!(true);
-    ensure_stream_usage(&mut payload);
-    let request = Request::builder()
-        .method(Method::POST)
-        .uri(uri)
-        .header(CONTENT_TYPE, "application/json")
-        .body(Full::new(HyperBytes::from(serde_json::to_vec(&payload)?)))?;
-    let start = Instant::now();
-    let response = client.request(request).await?;
-    let status = response.status();
-    if !status.is_success() {
-        let body = response.into_body().collect().await?.to_bytes();
-        let payload = serde_json::from_slice::<Value>(&body).unwrap_or_else(
-            |_| json!({"error": {"message": String::from_utf8_lossy(&body).trim().to_string()}}),
-        );
-        return Ok((payload, status));
-    }
-    let content_type = response
-        .headers()
-        .get(CONTENT_TYPE)
-        .and_then(|value| value.to_str().ok())
-        .unwrap_or("")
-        .to_ascii_lowercase();
-    if !content_type.contains("text/event-stream") {
-        let body = response.into_body().collect().await?.to_bytes();
-        let mut payload = serde_json::from_slice::<Value>(&body).unwrap_or_else(
-            |_| json!({"error": {"message": String::from_utf8_lossy(&body).trim().to_string()}}),
-        );
-        normalize_openai_usage(&mut payload);
-        return Ok((payload, status));
-    }
-    let mut aggregate = OpenAiStreamAggregate::new(response_model, start, "nonstream_via_stream");
-    let mut body = Body::new(response.into_body());
-    let mut buffered = Vec::<u8>::new();
-    while let Some(frame) = body.frame().await {
-        let frame = frame?;
-        let Some(data) = frame.data_ref() else {
-            continue;
-        };
-        buffered.extend_from_slice(data);
-        while let Some(index) = buffered.windows(2).position(|window| window == b"\n\n") {
-            let chunk = buffered.drain(..index + 2).collect::<Vec<_>>();
-            aggregate.process_sse_bytes(&chunk);
-        }
-    }
-    if !buffered.is_empty() {
-        aggregate.process_sse_bytes(&buffered);
-    }
-    let mut payload = aggregate.finish();
-    normalize_openai_usage(&mut payload);
-    Ok((payload, StatusCode::OK))
-}
-
-fn ensure_stream_usage(payload: &mut Value) {
-    let object = payload
-        .as_object_mut()
-        .expect("normalized chat payload should be an object");
-    let stream_options = object.entry("stream_options").or_insert_with(|| json!({}));
-    if !stream_options.is_object() {
-        *stream_options = json!({});
-    }
-    stream_options["include_usage"] = json!(true);
-}
-
-#[derive(Default)]
-struct AggregatedToolCall {
-    id: Option<String>,
-    kind: Option<String>,
-    name: Option<String>,
-    arguments: String,
-}
-
-struct OpenAiStreamAggregate {
-    metrics_mode: &'static str,
-    response_model: String,
-    started_at: Instant,
-    first_output_at: Option<Instant>,
-    id: Option<String>,
-    created: Option<u64>,
-    upstream_model: Option<String>,
-    system_fingerprint: Option<Value>,
-    role: Option<String>,
-    content: String,
-    content_truncated: bool,
-    reasoning_content: String,
-    reasoning_truncated: bool,
-    tool_calls: BTreeMap<u64, AggregatedToolCall>,
-    tool_arguments_truncated: bool,
-    finish_reason: Option<Value>,
-    usage: Option<Value>,
-}
-
-impl OpenAiStreamAggregate {
-    fn new(response_model: &str, started_at: Instant, metrics_mode: &'static str) -> Self {
-        Self {
-            metrics_mode,
-            response_model: response_model.to_string(),
-            started_at,
-            first_output_at: None,
-            id: None,
-            created: None,
-            upstream_model: None,
-            system_fingerprint: None,
-            role: None,
-            content: String::new(),
-            content_truncated: false,
-            reasoning_content: String::new(),
-            reasoning_truncated: false,
-            tool_calls: BTreeMap::new(),
-            tool_arguments_truncated: false,
-            finish_reason: None,
-            usage: None,
-        }
-    }
-
-    fn process_sse_bytes(&mut self, bytes: &[u8]) {
-        for event in parse_openai_sse_events(bytes) {
-            if let Ok(value) = serde_json::from_str::<Value>(&event) {
-                self.process_chunk(&value);
-            }
-        }
-    }
-
-    fn process_chunk(&mut self, chunk: &Value) {
-        if self.id.is_none() {
-            self.id = chunk.get("id").and_then(Value::as_str).map(str::to_string);
-        }
-        if self.created.is_none() {
-            self.created = chunk.get("created").and_then(Value::as_u64);
-        }
-        if self.upstream_model.is_none() {
-            self.upstream_model = chunk
-                .get("model")
-                .and_then(Value::as_str)
-                .map(str::to_string);
-        }
-        if self.system_fingerprint.is_none() {
-            self.system_fingerprint = chunk.get("system_fingerprint").cloned();
-        }
-        if let Some(usage) = chunk.get("usage")
-            && !usage.is_null()
-        {
-            self.usage = Some(usage.clone());
-        }
-        let Some(choice) = chunk
-            .get("choices")
-            .and_then(Value::as_array)
-            .and_then(|choices| choices.first())
-        else {
-            return;
-        };
-        if let Some(reason) = choice.get("finish_reason")
-            && !reason.is_null()
-        {
-            self.finish_reason = Some(reason.clone());
-        }
-        let delta = choice.get("delta").unwrap_or(&Value::Null);
-        if let Some(role) = delta.get("role").and_then(Value::as_str)
-            && self.role.is_none()
-        {
-            self.role = Some(role.to_string());
-        }
-        if let Some(content) = delta.get("content").and_then(Value::as_str)
-            && !content.is_empty()
-        {
-            self.mark_first_output();
-            self.content_truncated |= append_limited_stream_capture(&mut self.content, content);
-        }
-        for key in ["reasoning_content", "reasoning"] {
-            if let Some(reasoning) = delta.get(key).and_then(Value::as_str)
-                && !reasoning.is_empty()
-            {
-                self.mark_first_output();
-                self.reasoning_truncated |=
-                    append_limited_stream_capture(&mut self.reasoning_content, reasoning);
-            }
-        }
-        for tool_call in delta
-            .get("tool_calls")
-            .and_then(Value::as_array)
-            .into_iter()
-            .flatten()
-        {
-            self.mark_first_output();
-            let index = tool_call.get("index").and_then(Value::as_u64).unwrap_or(0);
-            let entry = self.tool_calls.entry(index).or_default();
-            if let Some(id) = tool_call.get("id").and_then(Value::as_str)
-                && !id.is_empty()
-            {
-                entry.id = Some(id.to_string());
-            }
-            if let Some(kind) = tool_call.get("type").and_then(Value::as_str)
-                && !kind.is_empty()
-            {
-                entry.kind = Some(kind.to_string());
-            }
-            let function = tool_call.get("function").unwrap_or(&Value::Null);
-            if let Some(name) = function.get("name").and_then(Value::as_str)
-                && !name.is_empty()
-            {
-                entry.name = Some(name.to_string());
-            }
-            if let Some(arguments) = function.get("arguments").and_then(Value::as_str)
-                && !arguments.is_empty()
-            {
-                self.tool_arguments_truncated |=
-                    append_limited_stream_capture(&mut entry.arguments, arguments);
-            }
-        }
-    }
-
-    fn finish(self) -> Value {
-        let ended_at = Instant::now();
-        let latency_ms = duration_ms(ended_at.duration_since(self.started_at));
-        let ttft_ms = self
-            .first_output_at
-            .map(|instant| duration_ms(instant.duration_since(self.started_at)));
-        let decode_ms = self
-            .first_output_at
-            .map(|instant| duration_ms(ended_at.duration_since(instant)));
-        let mut usage = self.usage.unwrap_or_else(|| json!({}));
-        normalize_openai_usage_object(&mut usage);
-        let observed = observed_metrics(&usage, latency_ms, ttft_ms, decode_ms);
-        let mut message = json!({
-            "role": self.role.unwrap_or_else(|| "assistant".to_string()),
-            "content": self.content,
-        });
-        if !self.reasoning_content.is_empty() {
-            message["reasoning_content"] = json!(self.reasoning_content);
-        }
-        let tool_calls = self
-            .tool_calls
-            .into_iter()
-            .map(|(index, tool)| {
-                json!({
-                    "index": index,
-                    "id": tool.id.unwrap_or_else(|| format!("call_{index}")),
-                    "type": tool.kind.unwrap_or_else(|| "function".to_string()),
-                    "function": {
-                        "name": tool.name.unwrap_or_default(),
-                        "arguments": tool.arguments,
-                    },
-                })
-            })
-            .collect::<Vec<_>>();
-        if !tool_calls.is_empty() {
-            message["tool_calls"] = Value::Array(tool_calls);
-        }
-        let response_truncated =
-            self.content_truncated || self.reasoning_truncated || self.tool_arguments_truncated;
-        let mut payload = json!({
-            "id": self.id.unwrap_or_else(make_chat_completion_id),
-            "object": "chat.completion",
-            "created": self.created.unwrap_or_else(unix_seconds),
-            "model": self.upstream_model.unwrap_or(self.response_model),
-            "choices": [{
-                "index": 0,
-                "message": message,
-                "finish_reason": self.finish_reason.unwrap_or(Value::String("stop".to_string())),
-            }],
-            "usage": usage,
-            "omniinfer_metrics": {
-                "mode": self.metrics_mode,
-                "latency_ms": latency_ms,
-                "ttft_ms": ttft_ms,
-                "decode_ms": decode_ms,
-                "observed_prefill_tps": observed.prefill_tps,
-                "observed_decode_tps": observed.decode_tps,
-                "response_truncated": response_truncated,
-            },
-        });
-        if let Some(fingerprint) = self.system_fingerprint {
-            payload["system_fingerprint"] = fingerprint;
-        }
-        payload
-    }
-
-    fn mark_first_output(&mut self) {
-        if self.first_output_at.is_none() {
-            self.first_output_at = Some(Instant::now());
-        }
-    }
-}
-
-fn append_limited_stream_capture(target: &mut String, text: &str) -> bool {
-    let current = target.chars().count();
-    if current >= MAX_STREAM_HISTORY_CAPTURE_CHARS {
-        return true;
-    }
-    let remaining = MAX_STREAM_HISTORY_CAPTURE_CHARS - current;
-    let incoming = text.chars().count();
-    if incoming <= remaining {
-        target.push_str(text);
-        return false;
-    }
-    target.extend(text.chars().take(remaining));
-    true
-}
-
-struct ObservedMetrics {
-    prefill_tps: Option<f64>,
-    decode_tps: Option<f64>,
-}
-
-fn observed_metrics(
-    usage: &Value,
-    _latency_ms: u64,
-    ttft_ms: Option<u64>,
-    decode_ms: Option<u64>,
-) -> ObservedMetrics {
-    let prompt_tokens = usage.get("prompt_tokens").and_then(Value::as_u64);
-    let completion_tokens = usage.get("completion_tokens").and_then(Value::as_u64);
-    ObservedMetrics {
-        prefill_tps: tokens_per_second(prompt_tokens, ttft_ms),
-        decode_tps: tokens_per_second(completion_tokens, decode_ms),
-    }
-}
-
-fn tokens_per_second(tokens: Option<u64>, millis: Option<u64>) -> Option<f64> {
-    let tokens = tokens?;
-    let millis = millis?;
-    if tokens == 0 || millis == 0 {
-        return None;
-    }
-    Some((tokens as f64) * 1000.0 / (millis as f64))
-}
-
-fn duration_ms(duration: Duration) -> u64 {
-    u64::try_from(duration.as_millis()).unwrap_or(u64::MAX)
-}
-
-fn unix_seconds() -> u64 {
-    SystemTime::now()
-        .duration_since(UNIX_EPOCH)
-        .map(|duration| duration.as_secs())
-        .unwrap_or(0)
-}
-
-fn make_chat_completion_id() -> String {
-    format!("chatcmpl-omniinfer-{}", unix_seconds())
-}
-
-fn normalize_openai_usage_object(usage: &mut Value) {
-    let Some(object) = usage.as_object_mut() else {
-        return;
-    };
-    if object.get("total_tokens").and_then(Value::as_u64).is_some() {
-        return;
-    }
-    let Some(prompt_tokens) = object.get("prompt_tokens").and_then(Value::as_u64) else {
-        return;
-    };
-    let Some(completion_tokens) = object.get("completion_tokens").and_then(Value::as_u64) else {
-        return;
-    };
-    object.insert(
-        "total_tokens".to_string(),
-        json!(prompt_tokens.saturating_add(completion_tokens)),
-    );
-}
-
-fn apply_proxy_model(payload: &mut Value, proxy_model: Option<&str>) {
-    if let Some(proxy_model) = proxy_model.filter(|value| !value.trim().is_empty()) {
-        payload["model"] = json!(proxy_model);
-    }
-}
-
+use openai_stream::{
+    StreamHistoryContext, apply_proxy_model, duration_ms, proxy_openai_nonstream_via_stream,
+    should_proxy_vllm_nonstream_via_stream, stream_openai_chat_with_history,
+};
 async fn clear_runtime_cache(
     client: &Client<HttpConnector, Full<HyperBytes>>,
     runtime_base: &str,
