+++ crates/omniinfer-cli/src/gateway/runtime_manager.rs
@@ -818,716 +818,18 @@ impl RustRuntimeManager {
     }
 }
 
-fn start_runtime_with_cold_start_policy(
-    backend_id: &str,
-    plan: &omniinfer_core::runtime_plan::ExternalRuntimePlan,
-    options: RuntimeProcessOptions,
-) -> Result<RuntimeProcess, RuntimeProcessError> {
-    let Some(initial_timeout) =
-        wsl_rocm_cold_start_retry_timeout(backend_id, options.startup_timeout)
-    else {
-        return RuntimeProcess::start(plan, options);
-    };
+mod lifecycle;
 
-    let total_timeout = options.startup_timeout;
-    retry_after_ready_timeout(
-        total_timeout,
-        initial_timeout,
-        WSL_ROCM_COLD_START_RETRY_COOLDOWN,
-        |attempt_timeout| {
-            let mut attempt_options = options.clone();
-            attempt_options.startup_timeout = attempt_timeout;
-            RuntimeProcess::start(plan, attempt_options)
-        },
-    )
-}
-
-fn wsl_rocm_cold_start_retry_timeout(
-    backend_id: &str,
-    total_timeout: Duration,
-) -> Option<Duration> {
-    (backend_id == "vllm-wsl2-rocm" && total_timeout >= WSL_ROCM_COLD_START_RETRY_MINIMUM_BUDGET)
-        .then_some(WSL_ROCM_COLD_START_INITIAL_ATTEMPT)
-}
-
-fn retry_after_ready_timeout<T>(
-    total_timeout: Duration,
-    initial_timeout: Duration,
-    cooldown: Duration,
-    mut attempt: impl FnMut(Duration) -> Result<T, RuntimeProcessError>,
-) -> Result<T, RuntimeProcessError> {
-    let started = Instant::now();
-    match attempt(initial_timeout) {
-        Err(RuntimeProcessError::ReadyTimeout) => {
-            let remaining_before_cooldown = total_timeout.saturating_sub(started.elapsed());
-            if remaining_before_cooldown <= cooldown {
-                return Err(RuntimeProcessError::ReadyTimeout);
-            }
-            eprintln!(
-                "OmniInfer: WSL2 ROCm cold start did not become ready after {} seconds; cooling down for {} seconds before retry",
-                initial_timeout.as_secs(),
-                cooldown.as_secs()
-            );
-            std::thread::sleep(cooldown);
-            let remaining = total_timeout.saturating_sub(started.elapsed());
-            if remaining.is_zero() {
-                return Err(RuntimeProcessError::ReadyTimeout);
-            }
-            eprintln!(
-                "OmniInfer: retrying WSL2 ROCm cold start once with the remaining {} seconds",
-                remaining.as_secs()
-            );
-            attempt(remaining)
-        }
-        result => result,
-    }
-}
-
-fn annotate_restore_state(
-    payload: &mut Value,
-    persistent_state: &local_state::LocalState,
-    loaded_runtimes: &BTreeMap<String, LoadedRustRuntime>,
-) {
-    let Some(selected) = persistent_state.selected_model.as_ref() else {
-        payload["restore_selection"] = Value::Null;
-        payload["restore_status"] = json!("not_configured");
-        payload["restore_completed"] = json!(false);
-        return;
-    };
-    let completed = loaded_runtimes.values().any(|loaded| {
-        loaded.route_state == RuntimeRouteState::Ready
-            && persistent_state
-                .selected_backend
-                .as_deref()
-                .is_none_or(|backend| loaded.backend_id == backend)
-            && loaded.model == selected.model
-            && loaded.mmproj == selected.mmproj
-            && loaded.ctx_size == selected.ctx_size
-            && loaded.request_defaults == selected.request_defaults
-    });
-    payload["restore_selection"] = json!({
-        "backend": persistent_state.selected_backend,
-        "model": selected.model,
-        "mmproj": selected.mmproj,
-        "ctx_size": selected.ctx_size,
-        "request_defaults": selected.request_defaults,
-    });
-    payload["restore_status"] = json!(if completed { "loaded" } else { "pending" });
-    payload["restore_completed"] = json!(completed);
-}
-
-fn same_load_configuration(
-    loaded: &LoadedRustRuntime,
-    backend_id: &str,
-    model_path: &str,
-    mmproj: Option<&str>,
-    ctx_size: Option<u32>,
-    launch_args: &[String],
-) -> bool {
-    loaded.backend_id == backend_id
-        && loaded.model == model_path
-        && loaded.mmproj.as_deref() == mmproj
-        && loaded.ctx_size == ctx_size
-        && loaded.launch_args == launch_args
-}
-
-fn model_load_response(loaded: &LoadedRustRuntime, already_loaded: bool) -> Value {
-    let info = loaded.process.info();
-    let mut response = json!({
-        "ok": true,
-        "already_loaded": already_loaded,
-        "requires_reload": false,
-        "model": loaded.model_key,
-        "owner_admin_id": loaded.owner_admin_id,
-        "selected_backend": loaded.backend_id,
-        "selected_model": loaded.model,
-        "selected_public_model_id": loaded.public_model_id,
-        "selected_mmproj": loaded.mmproj,
-        "selected_ctx_size": loaded.ctx_size,
-        "request_defaults": loaded.request_defaults,
-        "backend_pid": info.pid,
-        "backend_port": info.port,
-        "generation": loaded.generation,
-        "route_state": loaded.route_state.as_str(),
-        "allocation_id": loaded.allocation_id.get(),
-        "resource_budget": resource_budget_payload(&loaded.resource_budget),
-        "launch_command": info.command,
-        "log_path": info.log_path.display().to_string(),
-        "external_server_protocol": loaded.external_server_protocol.as_str(),
-        "client_endpoint": loaded.client_endpoint,
-        "openai_compatible": loaded.external_server_protocol.is_openai_compatible(),
-    });
-    if let Some(visible_devices) = loaded.cuda_visible_devices.as_deref() {
-        response["cuda_visible_devices"] = json!(visible_devices);
-    }
-    if let Some(warning) = loaded.cuda_warning.as_deref() {
-        response["warning"] = json!(warning);
-    }
-    response
-}
-
-struct RequestedRuntimeConfig<'a> {
-    backend_id: &'a str,
-    model_key: &'a str,
-    model_path: &'a str,
-    public_model_id: Option<&'a str>,
-    mmproj: Option<&'a str>,
-    ctx_size: Option<u32>,
-    request_defaults: &'a Map<String, Value>,
-    launch_args: &'a [String],
-}
-
-fn reload_required_response(
-    loaded: &LoadedRustRuntime,
-    requested: &RequestedRuntimeConfig<'_>,
-) -> Value {
-    json!({
-        "ok": false,
-        "already_loaded": true,
-        "requires_reload": true,
-        "error": {
-            "code": "model_reload_required",
-            "message": format!(
-                "model '{}' is already loaded with different runtime settings; unload it before selecting the new configuration",
-                requested.model_key,
-            ),
-        },
-        "current": {
-            "backend": loaded.backend_id,
-            "model": loaded.model_key,
-            "model_path": loaded.model,
-            "public_model_id": loaded.public_model_id,
-            "mmproj": loaded.mmproj,
-            "ctx_size": loaded.ctx_size,
-            "request_defaults": loaded.request_defaults,
-            "launch_args": loaded.launch_args,
-        },
-        "requested": {
-            "backend": requested.backend_id,
-            "model": requested.model_key,
-            "model_path": requested.model_path,
-            "public_model_id": requested.public_model_id,
-            "mmproj": requested.mmproj,
-            "ctx_size": requested.ctx_size,
-            "request_defaults": requested.request_defaults,
-            "launch_args": requested.launch_args,
-        },
-    })
-}
-
-fn loaded_runtime_payload(loaded: &LoadedRustRuntime) -> Value {
-    let info = loaded.process.info();
-    json!({
-        "id": loaded.model_key,
-        "owner_admin_id": loaded.owner_admin_id,
-        "backend": loaded.backend_id,
-        "model": loaded.model_key,
-        "model_path": loaded.model,
-        "public_model_id": loaded.public_model_id,
-        "mmproj": loaded.mmproj,
-        "ctx_size": loaded.ctx_size,
-        "request_defaults": loaded.request_defaults,
-        "runtime_mode": "external_server",
-        "backend_pid": info.pid,
-        "backend_port": info.port,
-        "generation": loaded.generation,
-        "route_state": loaded.route_state.as_str(),
-        "allocation_id": loaded.allocation_id.get(),
-        "resource_budget": resource_budget_payload(&loaded.resource_budget),
-        "launch_args": loaded.launch_args,
-        "cuda_visible_devices": loaded.cuda_visible_devices,
-        "warning": loaded.cuda_warning,
-        "launch_command": info.command,
-        "proxy_model": loaded.proxy_model_ref,
-        "external_server_protocol": loaded.external_server_protocol.as_str(),
-        "client_endpoint": loaded.client_endpoint,
-        "openai_compatible": loaded.external_server_protocol.is_openai_compatible(),
-        "backend_log": info.log_path.display().to_string(),
-    })
-}
-
+use lifecycle::*;
 const MIB: u64 = 1024 * 1024;
 const GIB: u64 = 1024 * MIB;
 
-fn build_runtime_resource_budget(
-    payload: &Value,
-    backend: &backend_registry::BackendSpec,
-    model: &str,
-    mmproj: Option<&str>,
-    ctx_size: u32,
-    cuda_visible_devices: Option<&str>,
-    replicate_across_domains: bool,
-) -> Result<ResourceBudget> {
-    let domains = if cfg!(target_os = "macos")
-        || backend
-            .capabilities
-            .iter()
-            .any(|value| value == "shared-memory")
-    {
-        vec![MemoryDomain::Unified("system".to_string())]
-    } else if backend.capabilities.iter().any(|value| value == "cuda") {
-        let devices = parse_cuda_devices(cuda_visible_devices.ok_or_else(|| {
-            anyhow::anyhow!("CUDA resource budgeting requires a selected device")
-        })?);
-        if devices.is_empty() {
-            anyhow::bail!("CUDA resource budgeting requires a selected device");
-        }
-        devices.into_iter().map(MemoryDomain::Cuda).collect()
-    } else {
-        vec![MemoryDomain::Host]
-    };
-    let explicit_total = payload
-        .get("resource_budget_bytes")
-        .and_then(Value::as_u64)
-        .filter(|bytes| *bytes > 0);
-    let weights = artifact_size_bytes(&PathBuf::from(model))?;
-    let projector = mmproj
-        .map(|path| artifact_size_bytes(&PathBuf::from(path)))
-        .transpose()?
-        .flatten()
-        .unwrap_or(0);
-    let Some(weights) = weights else {
-        let total = explicit_total.ok_or_else(|| {
-            anyhow::anyhow!(
-                "model artifact size is unknown; provide a non-zero resource_budget_bytes value"
-            )
-        })?;
-        return Ok(ResourceBudget::from_components(assign_component(
-            "client_provided_total",
-            total,
-            &domains,
-            replicate_across_domains,
-        )?)?);
-    };
-    let weights = weights.max(1);
-    let base = weights
-        .checked_add(projector)
-        .ok_or_else(|| anyhow::anyhow!("model artifact size overflow"))?;
-    let parameter_proxy = base.saturating_mul(2).max(GIB);
-    let ctx = u64::from(ctx_size.max(1));
-    let kv_cache = checked_scaled(parameter_proxy, 3, 100)?
-        .checked_mul(ctx)
-        .and_then(|bytes| bytes.checked_div(u64::from(DEFAULT_LOAD_CONTEXT_SIZE)))
-        .unwrap_or(u64::MAX)
-        .max(256 * MIB);
-    let activation_ctx = ctx.min(u64::from(DEFAULT_LOAD_CONTEXT_SIZE) * 4);
-    let activation = checked_scaled(parameter_proxy, 1, 100)?
-        .checked_mul(activation_ctx)
-        .and_then(|bytes| bytes.checked_div(u64::from(DEFAULT_LOAD_CONTEXT_SIZE)))
-        .unwrap_or(u64::MAX)
-        .max(128 * MIB);
-    let framework = checked_scaled(base, 8, 100)?.max(384 * MIB);
-    let allocator_slack = checked_scaled(base, 4, 100)?.max(160 * MIB);
-    let mut components = Vec::new();
-    for (name, bytes) in [
-        ("weights", weights),
-        ("kv_cache", kv_cache),
-        ("activation", activation),
-        ("framework_overhead", framework),
-        ("allocator_slack", allocator_slack),
-    ] {
-        components.extend(assign_component(
-            name,
-            bytes,
-            &domains,
-            replicate_across_domains,
-        )?);
-    }
-    if projector > 0 {
-        components.extend(assign_component(
-            "mmproj",
-            projector,
-            &domains,
-            replicate_across_domains,
-        )?);
-    }
-    let estimated = ResourceBudget::from_components(components)?;
-    if let Some(explicit_total) = explicit_total {
-        let estimated_minimum = if replicate_across_domains {
-            estimated.domains().values().copied().max().unwrap_or(0)
-        } else {
-            estimated
-                .domains()
-                .values()
-                .try_fold(0_u64, |total, bytes| total.checked_add(*bytes))
-                .ok_or_else(|| anyhow::anyhow!("resource budget overflow"))?
-        };
-        if explicit_total < estimated_minimum {
-            anyhow::bail!(
-                "resource_budget_bytes is below the estimated minimum of {estimated_minimum} bytes"
-            );
-        }
-        return Ok(ResourceBudget::from_components(assign_component(
-            "client_provided_total",
-            explicit_total,
-            &domains,
-            replicate_across_domains,
-        )?)?);
-    }
-    Ok(estimated)
-}
+mod resources;
 
-fn artifact_size_bytes(path: &PathBuf) -> Result<Option<u64>> {
-    let Ok(metadata) = fs::metadata(path) else {
-        return Ok(None);
-    };
-    if metadata.is_file() {
-        return Ok(Some(metadata.len()));
-    }
-    if !metadata.is_dir() {
-        return Ok(None);
-    }
-    let mut total = 0_u64;
-    let mut pending = vec![path.clone()];
-    while let Some(directory) = pending.pop() {
-        for entry in fs::read_dir(&directory)? {
-            let entry = entry?;
-            let file_type = entry.file_type()?;
-            if file_type.is_symlink() {
-                continue;
-            }
-            if file_type.is_dir() {
-                pending.push(entry.path());
-            } else if file_type.is_file() {
-                total = total
-                    .checked_add(entry.metadata()?.len())
-                    .ok_or_else(|| anyhow::anyhow!("model artifact size overflow"))?;
-            }
-        }
-    }
-    Ok((total > 0).then_some(total))
-}
+use resources::*;
+mod model_config;
 
-fn checked_scaled(value: u64, numerator: u64, denominator: u64) -> Result<u64> {
-    value
-        .checked_mul(numerator)
-        .and_then(|scaled| scaled.checked_div(denominator))
-        .ok_or_else(|| anyhow::anyhow!("resource budget overflow"))
-}
-
-fn assign_component(
-    name: &str,
-    bytes: u64,
-    domains: &[MemoryDomain],
-    replicate_across_domains: bool,
-) -> Result<Vec<BudgetComponent>> {
-    if replicate_across_domains {
-        if domains.is_empty() || bytes == 0 {
-            anyhow::bail!("resource component requires non-zero bytes and at least one domain");
-        }
-        return Ok(domains
-            .iter()
-            .map(|domain| BudgetComponent {
-                name: name.to_string(),
-                domain: domain.clone(),
-                bytes,
-            })
-            .collect());
-    }
-    distribute_component(name, bytes, domains)
-}
-
-fn distribute_component(
-    name: &str,
-    bytes: u64,
-    domains: &[MemoryDomain],
-) -> Result<Vec<BudgetComponent>> {
-    let count = u64::try_from(domains.len())?;
-    if count == 0 || bytes == 0 {
-        anyhow::bail!("resource component requires non-zero bytes and at least one domain");
-    }
-    let base = bytes / count;
-    let remainder = bytes % count;
-    domains
-        .iter()
-        .enumerate()
-        .map(|(index, domain)| {
-            let bytes = base + u64::from(u64::try_from(index)? < remainder);
-            Ok(BudgetComponent {
-                name: name.to_string(),
-                domain: domain.clone(),
-                bytes,
-            })
-        })
-        .collect()
-}
-
-fn parse_cuda_devices(visible_devices: &str) -> Vec<String> {
-    let mut devices = visible_devices
-        .split(',')
-        .map(str::trim)
-        .filter(|value| !value.is_empty())
-        .map(str::to_string)
-        .collect::<Vec<_>>();
-    devices.sort();
-    devices.dedup();
-    devices
-}
-
-fn detect_available_resources(
-    cuda_visible_devices: Option<&str>,
-) -> Result<BTreeMap<MemoryDomain, u64>> {
-    let mut system = sysinfo::System::new();
-    system.refresh_memory();
-    let available_memory = system.available_memory();
-    if available_memory == 0 {
-        anyhow::bail!("available system memory could not be detected");
-    }
-    let mut domains = BTreeMap::from([
-        (MemoryDomain::Host, available_memory),
-        (
-            MemoryDomain::Unified("system".to_string()),
-            available_memory,
-        ),
-    ]);
-    if let Some(devices) = cuda_visible_devices {
-        domains.extend(cuda_available_bytes(devices)?);
-    }
-    Ok(domains)
-}
-
-#[cfg(test)]
-fn detect_cuda_device_ids() -> Result<Vec<String>> {
-    Ok(vec!["0".to_string()])
-}
-
-#[cfg(not(test))]
-fn detect_cuda_device_ids() -> Result<Vec<String>> {
-    let output = std::process::Command::new("nvidia-smi")
-        .args(["--query-gpu=index", "--format=csv,noheader,nounits"])
-        .output()?;
-    if !output.status.success() {
-        anyhow::bail!("nvidia-smi device query failed");
-    }
-    let devices = parse_cuda_devices(&String::from_utf8_lossy(&output.stdout).replace('\n', ","));
-    if devices.is_empty() {
-        anyhow::bail!("nvidia-smi did not report any CUDA devices");
-    }
-    Ok(devices)
-}
-
-#[cfg(test)]
-fn cuda_available_bytes(visible_devices: &str) -> Result<BTreeMap<MemoryDomain, u64>> {
-    const TEST_CUDA_CAPACITY: u64 = 1024 * GIB;
-    let requested = parse_cuda_devices(visible_devices);
-    if requested.is_empty() {
-        anyhow::bail!("CUDA device selection is empty");
-    }
-    Ok(requested
-        .into_iter()
-        .map(|device| (MemoryDomain::Cuda(device), TEST_CUDA_CAPACITY))
-        .collect())
-}
-
-#[cfg(not(test))]
-fn cuda_available_bytes(visible_devices: &str) -> Result<BTreeMap<MemoryDomain, u64>> {
-    let requested = parse_cuda_devices(visible_devices);
-    if requested.is_empty() {
-        anyhow::bail!("CUDA device selection is empty");
-    }
-    let output = std::process::Command::new("nvidia-smi")
-        .args([
-            "--query-gpu=index,uuid,memory.free",
-            "--format=csv,noheader,nounits",
-        ])
-        .output()?;
-    if !output.status.success() {
-        anyhow::bail!("nvidia-smi memory query failed");
-    }
-    let rows = String::from_utf8_lossy(&output.stdout);
-    let mut available = BTreeMap::new();
-    for requested_device in &requested {
-        let memory_mib = rows.lines().find_map(|line| {
-            let parts = line.split(',').map(str::trim).collect::<Vec<_>>();
-            (parts.len() >= 3 && (parts[0] == *requested_device || parts[1] == *requested_device))
-                .then(|| parts[2].parse::<u64>().ok())
-                .flatten()
-        });
-        let memory_mib = memory_mib.ok_or_else(|| {
-            anyhow::anyhow!(
-                "selected CUDA device was not reported by nvidia-smi: {requested_device}"
-            )
-        })?;
-        available.insert(
-            MemoryDomain::Cuda(requested_device.clone()),
-            memory_mib
-                .checked_mul(MIB)
-                .ok_or_else(|| anyhow::anyhow!("CUDA capacity overflow"))?,
-        );
-    }
-    Ok(available)
-}
-
-fn merge_domain_totals(
-    left: &BTreeMap<MemoryDomain, u64>,
-    right: &BTreeMap<MemoryDomain, u64>,
-) -> Result<BTreeMap<MemoryDomain, u64>> {
-    let mut merged = left.clone();
-    for (domain, bytes) in right {
-        let total = merged.entry(domain.clone()).or_insert(0);
-        *total = total
-            .checked_add(*bytes)
-            .ok_or_else(|| anyhow::anyhow!("resource usage overflow"))?;
-    }
-    Ok(merged)
-}
-
-fn domain_bytes_payload(domains: &BTreeMap<MemoryDomain, u64>) -> Value {
-    Value::Object(
-        domains
-            .iter()
-            .map(|(domain, bytes)| (domain.key(), json!(bytes)))
-            .collect(),
-    )
-}
-
-fn resource_budget_payload(budget: &ResourceBudget) -> Value {
-    json!({
-        "domains_bytes": domain_bytes_payload(budget.domains()),
-        "components": budget.components().iter().map(|component| json!({
-            "name": component.name,
-            "domain": component.domain.key(),
-            "bytes": component.bytes,
-        })).collect::<Vec<_>>(),
-    })
-}
-
-fn model_log_file_name(base: &str, model_key: &str) -> String {
-    let sanitized = model_key
-        .chars()
-        .map(|ch| {
-            if ch.is_ascii_alphanumeric() || matches!(ch, '.' | '_' | '-') {
-                ch
-            } else {
-                '_'
-            }
-        })
-        .collect::<String>();
-    match base.rsplit_once('.') {
-        Some((stem, ext)) if !stem.is_empty() && !ext.is_empty() => {
-            format!("{stem}-{sanitized}.{ext}")
-        }
-        _ => format!("{base}-{sanitized}.log"),
-    }
-}
-
-fn json_required_str<'a>(payload: &'a Value, key: &'static str) -> Result<&'a str> {
-    payload
-        .get(key)
-        .and_then(Value::as_str)
-        .filter(|value| !value.trim().is_empty())
-        .ok_or_else(|| anyhow::anyhow!("field '{key}' is required"))
-}
-
-fn request_defaults_from_payload(payload: &Value) -> Result<Map<String, Value>> {
-    match payload.get("request_defaults") {
-        Some(Value::Object(defaults)) => Ok(defaults.clone()),
-        Some(_) => anyhow::bail!("request_defaults must be an object"),
-        None => Ok(Map::new()),
-    }
-}
-
-fn resolve_model_for_backend(
-    model: &str,
-    backend: &backend_registry::BackendSpec,
-) -> Result<omniinfer_core::model_artifacts::ResolvedModelArtifacts> {
-    if backend.model_artifact == "reference" {
-        return Ok(omniinfer_core::model_artifacts::ResolvedModelArtifacts {
-            model_path: model.to_string(),
-            mmproj_path: None,
-        });
-    }
-    let path = resolve_path_for_backend(model, backend, "model")?;
-    if backend.model_artifact == "vla-artifact" {
-        let path = PathBuf::from(&path);
-        if path.is_dir() {
-            anyhow::bail!(
-                "vla.cpp model must be a checkpoint file, not a directory: {}",
-                path.display()
-            );
-        }
-        if !is_vla_checkpoint_path(&path) {
-            anyhow::bail!(
-                "vla.cpp model must be a .gguf or .safetensors checkpoint: {}",
-                path.display()
-            );
-        }
-    }
-    if backend.model_artifact == "file" && PathBuf::from(&path).is_dir() {
-        return Ok(discover_llama_cpp_model_artifacts(&PathBuf::from(path))?);
-    }
-    Ok(omniinfer_core::model_artifacts::ResolvedModelArtifacts {
-        model_path: path,
-        mmproj_path: None,
-    })
-}
-
-fn is_vla_checkpoint_path(path: &std::path::Path) -> bool {
-    path.extension()
-        .and_then(|value| value.to_str())
-        .is_some_and(|extension| {
-            extension.eq_ignore_ascii_case("gguf") || extension.eq_ignore_ascii_case("safetensors")
-        })
-}
-
-fn resolve_path_for_backend(
-    text: &str,
-    backend: &backend_registry::BackendSpec,
-    label: &str,
-) -> Result<String> {
-    let mut path = expand_home(PathBuf::from(text.trim()));
-    if !path.is_absolute() {
-        let Some(models_dir) = backend.models_dir.as_deref() else {
-            anyhow::bail!("relative {label} path requires a configured models_dir");
-        };
-        path = PathBuf::from(models_dir).join(path);
-    }
-    if label == "model" && backend.model_artifact == "directory" {
-        if !path.is_dir() {
-            anyhow::bail!("model directory not found: {}", path.display());
-        }
-    } else if !path.exists() {
-        anyhow::bail!("{label} not found: {}", path.display());
-    }
-    Ok(path.display().to_string())
-}
-
-fn expand_home(path: PathBuf) -> PathBuf {
-    let text = path.to_string_lossy();
-    if let Some(rest) = text.strip_prefix("~/")
-        && let Some(home) = std::env::var_os("HOME")
-    {
-        return PathBuf::from(home).join(rest);
-    }
-    path
-}
-
-fn launch_args_have_ctx_size(family: &str, args: &[String]) -> bool {
-    args.iter().any(|arg| {
-        let flag = arg.split_once('=').map(|(flag, _)| flag).unwrap_or(arg);
-        match family {
-            "vllm" => flag == "--max-model-len",
-            "llama.cpp" | "turboquant" => matches!(flag, "-c" | "--ctx-size"),
-            _ => matches!(flag, "-c" | "--ctx-size" | "--max-model-len"),
-        }
-    })
-}
-
-fn merged_launch_args(
-    backend_id: &str,
-    family: &str,
-    defaults: &[String],
-    requested: Option<&[String]>,
-) -> Vec<String> {
-    let Some(requested) = requested else {
-        return defaults.to_vec();
-    };
-    if family != "llama.cpp" || !backend_id.starts_with("llama.cpp-") {
-        return requested.to_vec();
-    }
-    defaults.iter().chain(requested).cloned().collect()
-}
-
+use model_config::*;
 pub(super) fn pick_runtime_port(host: &str) -> Result<u16> {
     let listener = std::net::TcpListener::bind((host, 0))?;
     Ok(listener.local_addr()?.port())
