+++ crates/omniinfer-cli/src/gateway/runtime_manager/resources.rs
@@ -1,0 +1,344 @@
+use super::*;
+
+pub(super) fn build_runtime_resource_budget(
+    payload: &Value,
+    backend: &backend_registry::BackendSpec,
+    model: &str,
+    mmproj: Option<&str>,
+    ctx_size: u32,
+    cuda_visible_devices: Option<&str>,
+    replicate_across_domains: bool,
+) -> Result<ResourceBudget> {
+    let domains = if cfg!(target_os = "macos")
+        || backend
+            .capabilities
+            .iter()
+            .any(|value| value == "shared-memory")
+    {
+        vec![MemoryDomain::Unified("system".to_string())]
+    } else if backend.capabilities.iter().any(|value| value == "cuda") {
+        let devices = parse_cuda_devices(cuda_visible_devices.ok_or_else(|| {
+            anyhow::anyhow!("CUDA resource budgeting requires a selected device")
+        })?);
+        if devices.is_empty() {
+            anyhow::bail!("CUDA resource budgeting requires a selected device");
+        }
+        devices.into_iter().map(MemoryDomain::Cuda).collect()
+    } else {
+        vec![MemoryDomain::Host]
+    };
+    let explicit_total = payload
+        .get("resource_budget_bytes")
+        .and_then(Value::as_u64)
+        .filter(|bytes| *bytes > 0);
+    let weights = artifact_size_bytes(&PathBuf::from(model))?;
+    let projector = mmproj
+        .map(|path| artifact_size_bytes(&PathBuf::from(path)))
+        .transpose()?
+        .flatten()
+        .unwrap_or(0);
+    let Some(weights) = weights else {
+        let total = explicit_total.ok_or_else(|| {
+            anyhow::anyhow!(
+                "model artifact size is unknown; provide a non-zero resource_budget_bytes value"
+            )
+        })?;
+        return Ok(ResourceBudget::from_components(assign_component(
+            "client_provided_total",
+            total,
+            &domains,
+            replicate_across_domains,
+        )?)?);
+    };
+    let weights = weights.max(1);
+    let base = weights
+        .checked_add(projector)
+        .ok_or_else(|| anyhow::anyhow!("model artifact size overflow"))?;
+    let parameter_proxy = base.saturating_mul(2).max(GIB);
+    let ctx = u64::from(ctx_size.max(1));
+    let kv_cache = checked_scaled(parameter_proxy, 3, 100)?
+        .checked_mul(ctx)
+        .and_then(|bytes| bytes.checked_div(u64::from(DEFAULT_LOAD_CONTEXT_SIZE)))
+        .unwrap_or(u64::MAX)
+        .max(256 * MIB);
+    let activation_ctx = ctx.min(u64::from(DEFAULT_LOAD_CONTEXT_SIZE) * 4);
+    let activation = checked_scaled(parameter_proxy, 1, 100)?
+        .checked_mul(activation_ctx)
+        .and_then(|bytes| bytes.checked_div(u64::from(DEFAULT_LOAD_CONTEXT_SIZE)))
+        .unwrap_or(u64::MAX)
+        .max(128 * MIB);
+    let framework = checked_scaled(base, 8, 100)?.max(384 * MIB);
+    let allocator_slack = checked_scaled(base, 4, 100)?.max(160 * MIB);
+    let mut components = Vec::new();
+    for (name, bytes) in [
+        ("weights", weights),
+        ("kv_cache", kv_cache),
+        ("activation", activation),
+        ("framework_overhead", framework),
+        ("allocator_slack", allocator_slack),
+    ] {
+        components.extend(assign_component(
+            name,
+            bytes,
+            &domains,
+            replicate_across_domains,
+        )?);
+    }
+    if projector > 0 {
+        components.extend(assign_component(
+            "mmproj",
+            projector,
+            &domains,
+            replicate_across_domains,
+        )?);
+    }
+    let estimated = ResourceBudget::from_components(components)?;
+    if let Some(explicit_total) = explicit_total {
+        let estimated_minimum = if replicate_across_domains {
+            estimated.domains().values().copied().max().unwrap_or(0)
+        } else {
+            estimated
+                .domains()
+                .values()
+                .try_fold(0_u64, |total, bytes| total.checked_add(*bytes))
+                .ok_or_else(|| anyhow::anyhow!("resource budget overflow"))?
+        };
+        if explicit_total < estimated_minimum {
+            anyhow::bail!(
+                "resource_budget_bytes is below the estimated minimum of {estimated_minimum} bytes"
+            );
+        }
+        return Ok(ResourceBudget::from_components(assign_component(
+            "client_provided_total",
+            explicit_total,
+            &domains,
+            replicate_across_domains,
+        )?)?);
+    }
+    Ok(estimated)
+}
+
+pub(super) fn artifact_size_bytes(path: &PathBuf) -> Result<Option<u64>> {
+    let Ok(metadata) = fs::metadata(path) else {
+        return Ok(None);
+    };
+    if metadata.is_file() {
+        return Ok(Some(metadata.len()));
+    }
+    if !metadata.is_dir() {
+        return Ok(None);
+    }
+    let mut total = 0_u64;
+    let mut pending = vec![path.clone()];
+    while let Some(directory) = pending.pop() {
+        for entry in fs::read_dir(&directory)? {
+            let entry = entry?;
+            let file_type = entry.file_type()?;
+            if file_type.is_symlink() {
+                continue;
+            }
+            if file_type.is_dir() {
+                pending.push(entry.path());
+            } else if file_type.is_file() {
+                total = total
+                    .checked_add(entry.metadata()?.len())
+                    .ok_or_else(|| anyhow::anyhow!("model artifact size overflow"))?;
+            }
+        }
+    }
+    Ok((total > 0).then_some(total))
+}
+
+pub(super) fn checked_scaled(value: u64, numerator: u64, denominator: u64) -> Result<u64> {
+    value
+        .checked_mul(numerator)
+        .and_then(|scaled| scaled.checked_div(denominator))
+        .ok_or_else(|| anyhow::anyhow!("resource budget overflow"))
+}
+
+pub(super) fn assign_component(
+    name: &str,
+    bytes: u64,
+    domains: &[MemoryDomain],
+    replicate_across_domains: bool,
+) -> Result<Vec<BudgetComponent>> {
+    if replicate_across_domains {
+        if domains.is_empty() || bytes == 0 {
+            anyhow::bail!("resource component requires non-zero bytes and at least one domain");
+        }
+        return Ok(domains
+            .iter()
+            .map(|domain| BudgetComponent {
+                name: name.to_string(),
+                domain: domain.clone(),
+                bytes,
+            })
+            .collect());
+    }
+    distribute_component(name, bytes, domains)
+}
+
+pub(super) fn distribute_component(
+    name: &str,
+    bytes: u64,
+    domains: &[MemoryDomain],
+) -> Result<Vec<BudgetComponent>> {
+    let count = u64::try_from(domains.len())?;
+    if count == 0 || bytes == 0 {
+        anyhow::bail!("resource component requires non-zero bytes and at least one domain");
+    }
+    let base = bytes / count;
+    let remainder = bytes % count;
+    domains
+        .iter()
+        .enumerate()
+        .map(|(index, domain)| {
+            let bytes = base + u64::from(u64::try_from(index)? < remainder);
+            Ok(BudgetComponent {
+                name: name.to_string(),
+                domain: domain.clone(),
+                bytes,
+            })
+        })
+        .collect()
+}
+
+pub(super) fn parse_cuda_devices(visible_devices: &str) -> Vec<String> {
+    let mut devices = visible_devices
+        .split(',')
+        .map(str::trim)
+        .filter(|value| !value.is_empty())
+        .map(str::to_string)
+        .collect::<Vec<_>>();
+    devices.sort();
+    devices.dedup();
+    devices
+}
+
+pub(super) fn detect_available_resources(
+    cuda_visible_devices: Option<&str>,
+) -> Result<BTreeMap<MemoryDomain, u64>> {
+    let mut system = sysinfo::System::new();
+    system.refresh_memory();
+    let available_memory = system.available_memory();
+    if available_memory == 0 {
+        anyhow::bail!("available system memory could not be detected");
+    }
+    let mut domains = BTreeMap::from([
+        (MemoryDomain::Host, available_memory),
+        (
+            MemoryDomain::Unified("system".to_string()),
+            available_memory,
+        ),
+    ]);
+    if let Some(devices) = cuda_visible_devices {
+        domains.extend(cuda_available_bytes(devices)?);
+    }
+    Ok(domains)
+}
+
+#[cfg(test)]
+pub(super) fn detect_cuda_device_ids() -> Result<Vec<String>> {
+    Ok(vec!["0".to_string()])
+}
+
+#[cfg(not(test))]
+pub(super) fn detect_cuda_device_ids() -> Result<Vec<String>> {
+    let output = std::process::Command::new("nvidia-smi")
+        .args(["--query-gpu=index", "--format=csv,noheader,nounits"])
+        .output()?;
+    if !output.status.success() {
+        anyhow::bail!("nvidia-smi device query failed");
+    }
+    let devices = parse_cuda_devices(&String::from_utf8_lossy(&output.stdout).replace('\n', ","));
+    if devices.is_empty() {
+        anyhow::bail!("nvidia-smi did not report any CUDA devices");
+    }
+    Ok(devices)
+}
+
+#[cfg(test)]
+pub(super) fn cuda_available_bytes(visible_devices: &str) -> Result<BTreeMap<MemoryDomain, u64>> {
+    const TEST_CUDA_CAPACITY: u64 = 1024 * GIB;
+    let requested = parse_cuda_devices(visible_devices);
+    if requested.is_empty() {
+        anyhow::bail!("CUDA device selection is empty");
+    }
+    Ok(requested
+        .into_iter()
+        .map(|device| (MemoryDomain::Cuda(device), TEST_CUDA_CAPACITY))
+        .collect())
+}
+
+#[cfg(not(test))]
+pub(super) fn cuda_available_bytes(visible_devices: &str) -> Result<BTreeMap<MemoryDomain, u64>> {
+    let requested = parse_cuda_devices(visible_devices);
+    if requested.is_empty() {
+        anyhow::bail!("CUDA device selection is empty");
+    }
+    let output = std::process::Command::new("nvidia-smi")
+        .args([
+            "--query-gpu=index,uuid,memory.free",
+            "--format=csv,noheader,nounits",
+        ])
+        .output()?;
+    if !output.status.success() {
+        anyhow::bail!("nvidia-smi memory query failed");
+    }
+    let rows = String::from_utf8_lossy(&output.stdout);
+    let mut available = BTreeMap::new();
+    for requested_device in &requested {
+        let memory_mib = rows.lines().find_map(|line| {
+            let parts = line.split(',').map(str::trim).collect::<Vec<_>>();
+            (parts.len() >= 3 && (parts[0] == *requested_device || parts[1] == *requested_device))
+                .then(|| parts[2].parse::<u64>().ok())
+                .flatten()
+        });
+        let memory_mib = memory_mib.ok_or_else(|| {
+            anyhow::anyhow!(
+                "selected CUDA device was not reported by nvidia-smi: {requested_device}"
+            )
+        })?;
+        available.insert(
+            MemoryDomain::Cuda(requested_device.clone()),
+            memory_mib
+                .checked_mul(MIB)
+                .ok_or_else(|| anyhow::anyhow!("CUDA capacity overflow"))?,
+        );
+    }
+    Ok(available)
+}
+
+pub(super) fn merge_domain_totals(
+    left: &BTreeMap<MemoryDomain, u64>,
+    right: &BTreeMap<MemoryDomain, u64>,
+) -> Result<BTreeMap<MemoryDomain, u64>> {
+    let mut merged = left.clone();
+    for (domain, bytes) in right {
+        let total = merged.entry(domain.clone()).or_insert(0);
+        *total = total
+            .checked_add(*bytes)
+            .ok_or_else(|| anyhow::anyhow!("resource usage overflow"))?;
+    }
+    Ok(merged)
+}
+
+pub(super) fn domain_bytes_payload(domains: &BTreeMap<MemoryDomain, u64>) -> Value {
+    Value::Object(
+        domains
+            .iter()
+            .map(|(domain, bytes)| (domain.key(), json!(bytes)))
+            .collect(),
+    )
+}
+
+pub(super) fn resource_budget_payload(budget: &ResourceBudget) -> Value {
+    json!({
+        "domains_bytes": domain_bytes_payload(budget.domains()),
+        "components": budget.components().iter().map(|component| json!({
+            "name": component.name,
+            "domain": component.domain.key(),
+            "bytes": component.bytes,
+        })).collect::<Vec<_>>(),
+    })
+}
